+++ swing/framework-ui/src/main/java/is/codion/swing/framework/ui/EntityApplicationPanel.java
@@ -297,22 +297,20 @@ public abstract class EntityApplicationPanel<M extends SwingEntityApplicationMod
 	}
 
 	/**
-	 * @param <T> the layout type
 	 * @return the application layout
 	 */
-	public final <T extends ApplicationLayout> T applicationLayout() {
-		return (T) applicationLayout;
+	public final ApplicationLayout applicationLayout() {
+		return applicationLayout;
 	}
 
 	/**
-	 * @param <T> the entity panel type
 	 * @param entityType the entityType
 	 * @return the first entity panel found based on the given entity type
 	 * @throws IllegalArgumentException in case this application panel does not contain a panel for the given entity type
 	 */
-	public final <T extends EntityPanel> T entityPanel(EntityType entityType) {
+	public final EntityPanel entityPanel(EntityType entityType) {
 		requireNonNull(entityType);
-		return (T) entityPanels.stream()
+		return entityPanels.stream()
 						.filter(entityPanel -> entityPanel.model().entityType().equals(entityType))
 						.findFirst()
 						.orElseThrow(() -> new IllegalArgumentException("EntityPanel for entity: " + entityType + " not found in application panel: " + getClass()));