+++ swing/framework-ui/src/main/java/is/codion/swing/framework/ui/EntityEditComponentPanel.java
@@ -170,11 +170,10 @@ public class EntityEditComponentPanel extends JPanel {
 	}
 
 	/**
-	 * @param <T> the edit model type
 	 * @return the edit model this panel is based on
 	 */
-	public final <T extends SwingEntityEditModel> T editModel() {
-		return (T) editModel;
+	public final SwingEntityEditModel editModel() {
+		return editModel;
 	}
 
 	/**